+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## VNext
 - [Upgrade System.Diagnostics.PerformanceCounter to version 6.0.0 to address CVE-2021-24112](https://github.com/microsoft/ApplicationInsights-dotnet/pull/2707)
+- [Cache authentication token to improve performance when using AAD authentication](https://github.com/microsoft/ApplicationInsights-dotnet/issues/2539)
 
 ## Version 2.22.0-beta1
 - Update endpoint redirect header name for QuickPulse module to v2